Move most files out of `src/plugins/`
