+++ plugins/phaser/cache-busted-loader-plugin.ts
