+++ plugins/typedoc/typedoc-plugin-rename-svg.js
