+++ plugins/vite/vite-minify-json-plugin.ts
